aider: Add AssetOverviewPage to the self.frames dictionary in PortfolioTrackerApp __init__ method.
# Aider chat conversation:

USER: --help

ASSISTANT: Ok.

USER: when I run the program I get an error. Exception in Tkinter callback
Traceback (most recent call last):
  File "/Users/psps/anaconda3/envs/portfoliotracker_env/lib/python3.11/tkinter/__init__.py", line 1948, in __call__
    return self.func(*args)
           ^^^^^^^^^^^^^^^^
  File "/Users/psps/gits/PortfolioTracker/ui.py", line 71, in <lambda>
    command=lambda: self.show_frame(AssetOverviewPage))
                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/psps/gits/PortfolioTracker/ui.py", line 77, in show_frame
    frame = self.frames[cont]
            ~~~~~~~~~~~^^^^^^
KeyError: <class 'ui.AssetOverviewPage'>

ASSISTANT: I'm sorry to hear that you're experiencing an error. The error message indicates that the `AssetOverviewPage` class is not found in the `self.frames` dictionary when the `show_frame` method is called.

To help you further, I need to see the code where the `self.frames` dictionary is populated. This is typically done in the `__init__` method of the `PortfolioTrackerApp` class in the `ui.py` file.

Could you please add the `ui.py` file to the chat so I can take a closer look?

USER: I added these *read-write* files: ui.py

ASSISTANT: The error is due to the `AssetOverviewPage` not being included in the `self.frames` dictionary. This dictionary is used to store all the pages (frames) of the application. When you try to show the `AssetOverviewPage`, it can't be found in the dictionary, hence the `KeyError`.

To fix this, we need to add `AssetOverviewPage` to the `self.frames` dictionary in the `__init__` method of the `PortfolioTrackerApp` class.

Here is the *SEARCH/REPLACE* block:

ui.py
```python
<<<<<<< SEARCH
for F in (TransactionsPage, NetValuePage):
=======
for F in (TransactionsPage, NetValuePage, AssetOverviewPage):
>>>>>>> REPLACE
    frame = F(container, self, self.db_file)
    self.frames[F] = frame
    frame.grid(row=0, column=0, sticky="nsew")
```

This change will add `AssetOverviewPage` to the `self.frames` dictionary, so it can be found when you try to show it.